da3_mcp/requirements.txt
mcp[cli]>=1.2.0
rawpy>=0.18
imageio>=2.28
piexif>=1.1
Pillow>=10.0
rasterio>=1.3
pyproj>=3.5
scipy>=1.11
scikit-learn>=1.3
numpy>=1.24
